+++ slips_files/core/database.py
@@ -1327,8 +1327,8 @@ class Database(object):
         if current_data:
             if 'asn' in current_data.keys():
                 asn = current_data['asn']['asnorg']
-                if 'Unknown' not in asn:
-                    identification += 'AS: ' + current_data['asn']['asnorg'] + ', '
+                if 'Unknown' not in asn and asn != '':
+                    identification += 'AS: ' + asn + ', '
             if 'SNI' in current_data.keys():
                 SNI = current_data['SNI']
                 if type(SNI) == list:
@@ -2117,6 +2117,12 @@ class Database(object):
 
         return gateway
 
+    def get_ssl_info(self, sha1):
+        info = self.rcache.hmget('IoC_SSL', sha1)[0]
+        if info == None:
+            return False
+        return info
+
     def set_profile_module_label(self, profileid, module, label):
         """
         Set a module label for a profile.
@@ -2187,11 +2193,20 @@ class Database(object):
     def add_ja3_to_IoC(self, ja3_dict) -> None:
         """
         Store a group of ja3 in the db
-        :param ja3_dict: a json serialized dict {ja3: {'source':..,'tags':..,
-                                                        'threat_level':... ,'description'}}
+        :param ja3_dict:  {ja3: {'source':..,'tags':..,
+                            'threat_level':... ,'description'}}
 
         """
         self.rcache.hmset('IoC_JA3', ja3_dict)
+
+    def add_ssl_sha1_to_IoC(self, malicious_ssl_certs):
+        """
+        Store a group of ssl fingerprints in the db
+        :param malicious_ssl_certs:  {sha1: {'source':..,'tags':..,
+                                    'threat_level':... ,'description'}}
+
+        """
+        self.rcache.hmset('IoC_SSL', malicious_ssl_certs)
 
     def add_ip_to_IoC(self, ip: str, description: str) -> None:
         """